Rename .es6 files to .js
