+++ app/javascript/legacy/bank_accounts/confirm/index.js
