+++ app/javascript/legacy/bank_accounts/confirm/page.js
@@ -1,2 +1,2 @@
 // License: LGPL-3.0-or-later
-require('./index.es6')
+require('./')
