+++ app/javascript/legacy/bank_accounts/create.js
