+++ app/javascript/legacy/campaigns/show/page.js
@@ -35,7 +35,7 @@ if(app.current_campaign_editor) {
 	require('./admin')
 	appl.def('current_campaign_editor', true)
 	require('../../gift_options/admin')
-	var create_info_card = require('../../supporters/info-card.es6')
+	var create_info_card = require('../../supporters/info-card.js')
 }
 
 // Initialize the state for the top-level campaign component
