+++ app/javascript/legacy/campaigns/supporters/index/metrics.js
