+++ app/javascript/legacy/campaigns/supporters/index/page.js
@@ -1,5 +1,5 @@
 // License: LGPL-3.0-or-later
 require('../../timeline')
 require('../../totals')
-require('./index.es6')
+require('./')
