+++ app/javascript/legacy/common/direct-to-s3-upload.js
