+++ app/javascript/legacy/common/editor/froala.js
