+++ app/javascript/legacy/common/editor/quill.js
