+++ app/javascript/legacy/components/color-picker.js
