+++ app/javascript/legacy/components/show-more-button.js
