+++ app/javascript/legacy/components/tables/search.js
