+++ app/javascript/legacy/nonprofits/recurring_donations/index/index.js
