+++ app/javascript/legacy/nonprofits/supporter_form/page.js
@@ -1,2 +1,2 @@
 // License: LGPL-3.0-or-later
-require('./index.es6')
+require('./')
