+++ app/javascript/legacy/recurring_donations/edit/page.js
@@ -1,2 +1,2 @@
 // License: LGPL-3.0-or-later
-require("./index.es6")
+require("./")
