+++ app/javascript/legacy/recurring_donations/edit/payment-step.js
@@ -5,7 +5,7 @@ const flyd = require('flyd')
 flyd.lift = require('flyd/module/lift')
 flyd.flatMap = require('flyd/module/flatmap')
 const request = require('../../common/request')
-const cardForm = require('./card-form.es6')
+const cardForm = require('./card-form')
 const format = require('../../common/format').default
 const progressBar = require('../../components/progress-bar')
 const utils = require('../../common/utilities').default
