+++ app/javascript/legacy/settings/index/page.js
@@ -93,7 +93,7 @@ appl.def('update_change_amount_message', function(node) {
 
 if(app.current_nonprofit_user) {
 	appl.verify_identity = require('../../nonprofits/payouts/index/verify_identity')
-	appl.create_bank_account = require('../../bank_accounts/create.es6')
+	appl.create_bank_account = require('../../bank_accounts/create')
 }
 
 appl.def('statement.validate', function(node) {
