+++ app/javascript/legacy/supporters/info-card.js
